all tests passinggit add .! integration complete?
