--- compounds/tests/test_registry.py
@@ -1,4 +1,0 @@
-from compounds.CompoundDB import db
-
-
-print(db.get_compound_names())
